Added page specific decorators
