+++ blocks/page/page.js
@@ -1,30 +1,6 @@
-import { convertToBackground, decorateLink, decorateTagLink, processDivisions, normalizePropertyValue, normalizePropertyName } from "../../scripts/helpers.js";
+import { processDivisions, normalizePropertyName } from "../../scripts/helpers.js";
+import { PagePropertiesController } from "../../scripts/page-properties.js";
 
-const PagePropertiesController = {
-    callbacks: [],
-    ready: false,
-    properties: null,
-    setProperties(props) {
-        PagePropertiesController.properties = props;
-        PagePropertiesController.ready = true;
-        console.log("Page Properties:", this.properties);
-        PagePropertiesController.runCallbacks();
-    },
-    runCallbacks() {
-        while ( PagePropertiesController.callbacks.length > 0) {
-            const cb = PagePropertiesController.callbacks.shift();
-            cb(PagePropertiesController.properties);
-        }
-    },
-};
-export const resolvePageProperties = function resolvePageProperties(callback) {
-    if (PagePropertiesController.ready) {
-        callback(PagePropertiesController.properties);
-    } else {
-        PagePropertiesController.callbacks.push(callback);
-    }
-};
-window.resolvePageProperties = resolvePageProperties;
 
 function applyPathClassesToPage({ name }) {
     const parts = location.pathname.split("/");
