+++ scripts/pages/article.js
@@ -3,3 +3,12 @@
  *   All rights reserved.
  *   This file, its contents, concepts, methods, behavior, and operation  (collectively the "Software") are protected by trade secret, patent,  and copyright laws. The use of the Software is governed by a license  agreement. Disclosure of the Software to third parties, in any form,  in whole or in part, is expressly prohibited except as authorized by the license agreement.
  */
+
+export default function decorate($main) {
+    var paragraphs = document.querySelectorAll("body > main > div > div > p");
+    paragraphs.forEach(p => {
+      if( p.querySelector("picture") ) {
+        p.classList.add("article-picture");
+      }
+    });
+}
